refactor(FE): `useCreateGym` hook
