+++ frontend/src/components/Gyms/Gyms.js
@@ -2,8 +2,8 @@ import React from "react";
 import Button from "react-bootstrap/Button";
 import Card from "react-bootstrap/Card";
 import { Link } from "react-router-dom";
-import { useCreateGym } from "../../hooks/mutations";
 import useAuth from "../../hooks/useAuth";
+import useCreateGym from "../../hooks/useCreateGym";
 import useFakemonsByUser from "../../hooks/useFakemonsByUser";
 import useGyms from "../../hooks/useGyms";
 import useIsRegistered from "../../hooks/useIsRegistered";
